Db setup #1

See merge request courses/2020-fall/csci-5308/group7/csci5308!30

+++ src/main/java/data/ILeagueFactory.java
@@ -4,7 +4,7 @@ import model.League;
 
 public interface ILeagueFactory {
 
-    long addLeague(League league) throws Exception;
-    void loadLeagueByName(long id, League league) throws Exception;
+    int addLeague(League league) throws Exception;
+    void loadLeagueByName(int id, League league) throws Exception;
 
 }
